# packages/contracts/generated/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/sync/conflicts/{id}": {
      "patch": {
        "operationId": "SyncController_resolveConflict",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["Sync"]
      }
    }
  },
  "info": {
    "title": "DAJA Platform API",
    "description": "Versioned DAJA Platform HTTP API contract generated from Nest controllers.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
...
}